+++ cah/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = '2.2.5'
-__update_date__ = '2024-01-16_19:01:36'
+__version__ = '2.2.6'
+__update_date__ = '2024-01-17_08:42:11'
 from pathlib import Path
 
 ROOT_PATH = Path(__file__)